requirements.txt
streamlit==1.34.0
opencv-python==4.10.0.84
scipy==1.14.0
numpy>=1.18.5
tqdm>=4.64.0
Pillow>=7.1.2
gdown==5.2.0
htbuilder==0.6.2
torchtext==0.15.1
torchvision==0.16.0
torchaudio==2.1.0
torch==2.1.0
transformers==4.27.1